Added methods to help debugging the code
