+++ src/main/java/pl/szczepanik/silencio/core/Value.java
@@ -30,4 +30,9 @@ public class Value {
     public Object getValue() {
         return value;
     }
+
+    @Override
+    public String toString() {
+        return String.format("Value [value='%s']", value);
+    }
 }
