+++ src/test/java/pl/szczepanik/silencio/core/KeyTest.java
@@ -23,4 +23,18 @@ public class KeyTest extends GenericTest {
         // then
         assertThat(key.getKey()).isEqualTo(keyString);
     }
+
+    @Test
+    public void shouldReturnValidFormat() {
+
+        // given
+        String keyName = "golden key";
+        Key key = new Key(keyName);
+
+        // when
+        String dump = key.toString();
+
+        // then
+        assertThat(dump).isEqualTo(String.format("Key [key='%s']", keyName));
+    }
 }
